Renamed variables, functions, filenames.
